flask_app/pyproject.toml
[tool.poetry]
name = "flask-app"
version = "0.1.0"
description = ""
authors = ["Sean Deery <sdeery14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
